# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v1"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Channel","description":"Channel operations"}],"paths":{"/peertube/v1/{channelId}":{"get":{"tags":["Channel"],"summary":"Retrieve a channel by its channel id","description":"Get a channelDTO object by specifying its channel id","operationId":"getChannel","parameters":[{"name":"channelId","in":"path","description":"Name of channel to be searched","required":true,"schema":{"type":"string"},"example":"transport_evolved_take_2@peertube.tv"},{"name":"maxVideos","in":"query","description":"Maximum number of videos to be searched","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"maxComments","in":"query","description":"Maximum number of comments per video to be searched","required":false,"schema":{"type":"integer","format":"int32","default":2}}],"responses":{"200":{"description":"List of channels","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDTO"},"examples":{"Channel example":{"summary":"Successful response","description":"This example provides the representation of a valid response","value":{"id":"80","name":"tv","description":"canal de videos random","createdTime":"2023-01-01T23:36:17.306Z","videos":[{"id":"21856","name":"Alien: Isolation live on iOS - 1/9/2023, 10:17:06 PM","description":"HOW WILL YOU SURVIVE? Discover the true meaning of fear in Alien: Isolation, a survival horror set in an atmosphere of constant dread and mortal danger.","releaseTime":"2023-01-09T22:32:41.126Z","comments":[{"id":"3955","text":"Funniest video I've ever seen","createdOn":"2023-01-03T23:45:40.196Z"}],"captions":[{"id":"1","link":"/lazy-static/video-captions/68efe0a2-8ed8-4a6d-831a-ab8f5fc8f7fc-cy.vtt","language":"Spanish"}],"user":{"name":"stux","user_link":"https://peertube.tv/accounts/stux","picture_link":"https://peertube.tv/lazy-static/avatars/f8463b3f-2e2e-4f36-8bd3-fa9c0ef4f463.png"}}]}}}}}},"404":{"description":"❌ **Channel not found**"}}},"post":{"tags":["Channel"],"summary":"Save a channel by its channel id","description":"Post a channelDTO object to VideoMiner by specifying its channel id","operationId":"postChannel","parameters":[{"name":"channelId","in":"path","description":"Name of channel to be created","required":true,"schema":{"type":"string"}},{"name":"maxVideos","in":"query","description":"Maximum number of videos to be included","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"maxComments","in":"query","description":"Maximum number of comments per video to be included","required":false,"schema":{"type":"integer","format":"int32","default":2}}],"responses":{"201":{"description":"Channel successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDTO"}}}},"500":{"description":"The sent channel object  was incorrectly formed. Usually caused by a non existing channel name"}}}}},"components":{"schemas":{"CaptionDTO":{"type":"object","properties":{"id":{"type":"string"},"link":{"type":"string"},"language":{"type":"string"}}},"ChannelDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"createdTime":{"type":"string"},"videos":{"type":"array","items":{"$ref":"#/components/schemas/VideosDTO"}}}},"CommentDTO":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string"},"createdOn":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"user_link":{"type":"string"},"picture_link":{"type":"string"}}},"VideosDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"releaseTime":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDTO"},"captions":{"type":"array","items":{"$ref":"#/components/schemas/CaptionDTO"}},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentDTO"}}}}}}}